Divide MeshIntegrator classes
